Create stackPractice_01.c
This was a test to see if I still remember a C stack implementation with some terminal UI as a bonus. Looks like I missed the "makeEmpty()" function. It will serve as great practice with exception handling & input validation in the future.

Estimated time to write: 35 minutes (would probably be 20 minutes without making the UI)